move IpUtils and IpUtilsTest to soul-common, and use it soul-admin
