+++ soul-admin/src/test/java/org/dromara/soul/admin/listener/ApplicationStartListenerTest.java
@@ -17,15 +17,13 @@
 
 package org.dromara.soul.admin.listener;
 
+import org.dromara.soul.common.utils.IpUtils;
 import org.junit.Test;
 import org.junit.runner.RunWith;
 import org.springframework.boot.test.context.SpringBootTest;
 import org.springframework.boot.web.server.LocalServerPort;
 import org.springframework.test.context.ActiveProfiles;
 import org.springframework.test.context.junit4.SpringJUnit4ClassRunner;
-
-import java.net.InetAddress;
-import java.net.UnknownHostException;
 
 import static org.junit.Assert.assertEquals;
 
@@ -44,12 +42,7 @@ public final class ApplicationStartListenerTest {
 
     @Test
     public void testOnApplicationEvent() {
-        String host;
-        try {
-            host = InetAddress.getLocalHost().getHostAddress();
-        } catch (UnknownHostException e) {
-            host = "127.0.0.1";
-        }
+        String host = IpUtils.getHost();
         String expectedPath = "http://" + String.join(":", host, String.valueOf(port));
         assertEquals(expectedPath, SoulDomain.getInstance().getHttpPath());
     }
