+++ soul-common/src/test/java/org/dromara/soul/common/utils/IpUtilsTest.java
@@ -15,7 +15,7 @@
  * limitations under the License.
  */
 
-package org.dromara.soul.client.springmvc.utils;
+package org.dromara.soul.common.utils;
 
 import org.junit.After;
 import org.junit.Before;
@@ -24,6 +24,7 @@ import org.mockito.MockedStatic;
 
 import java.net.InetAddress;
 import java.net.UnknownHostException;
+
 import static org.junit.Assert.assertEquals;
 import static org.mockito.Mockito.mock;
 import static org.mockito.Mockito.mockStatic;
